+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-For a full diff see [`1.14.1...main`][1.14.1...main].
+For a full diff see [`1.15.0...main`][1.15.0...main].
+
+## [`1.15.0`][1.15.0]
+
+For a full diff see [`1.14.1...1.15.0`][1.14.1...1.15.0].
 
 ### Added
 
@@ -315,6 +319,7 @@ For a full diff see [`fd198f0...0.1.0`][fd198f0...0.1.0].
 [1.13.2]: https://github.com/ergebnis/rector-rules/releases/tag/1.13.2
 [1.14.0]: https://github.com/ergebnis/rector-rules/releases/tag/1.14.0
 [1.14.1]: https://github.com/ergebnis/rector-rules/releases/tag/1.14.1
+[1.15.0]: https://github.com/ergebnis/rector-rules/releases/tag/1.15.0
 
 [fd198f0...0.1.0]: https://github.com/ergebnis/rector-rules/compare/fd198f0...0.1.0
 [0.1.0...0.2.0]: https://github.com/ergebnis/rector-rules/compare/0.1.0...0.2.0
@@ -346,7 +351,8 @@ For a full diff see [`fd198f0...0.1.0`][fd198f0...0.1.0].
 [1.13.1...1.13.2]: https://github.com/ergebnis/rector-rules/compare/1.13.1...1.13.2
 [1.13.2...1.14.0]: https://github.com/ergebnis/rector-rules/compare/1.13.2...1.14.0
 [1.14.0...1.14.1]: https://github.com/ergebnis/rector-rules/compare/1.14.0...1.14.1
-[1.14.1...main]: https://github.com/ergebnis/rector-rules/compare/1.14.1...main
+[1.14.1...1.15.0]: https://github.com/ergebnis/rector-rules/compare/1.14.1...1.15.0
+[1.15.0...main]: https://github.com/ergebnis/rector-rules/compare/1.15.0...main
 
 [#1]: https://github.com/ergebnis/rector-rules/pull/1
 [#3]: https://github.com/ergebnis/rector-rules/pull/3